requirements.txt
numpy
pandas
matplotlib
qiskit==0.46.3
qiskit-aer-gpu==0.13.3
qiskit-terra==0.46.3
